Interim solution to string pointers
This is a bit of a hack, but it makes `Fiddle::Pointer`s to strings "act" like they do in MRI, somewhat.

The problem:
MRI pointers _actually_ point to the underlying string's memory, whereas JRuby pointers malloc and copy the string's contents. This is because, in Java, we cannot gain direct access to the string's memory. The result is that, while in MRI you can change a string by modifying the bytes through a Pointer, in JRuby, modifying the bytes has no effect on the original string.

The Solution:
We do the malloc/copy, but keep a reference to the passed in string. Any change to the pointer's memory is then performed on the string, and any time the pointer's memory is read, we re-copy the contents of the string beforehand. We also perform these workarounds when passing the string pointer to a `Fiddle::Function`, since functions like strcpy can modify the string.

While this is not the most ideal solution (and certainly not the most performant), it does make things a little less astonishing most of the time, and should serve its purpose in the time being. Developers making heavy use of string pointers and are concerned about performcance should manually malloc/copy their own Pointers for now, which should be cross-ruby compatible at this point.